requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
Markdown==3.7
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
sslcommerz-lib==1.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
